requirements.txt
humanfriendly
pyyaml
distorm3
-e git+git@github.com:dwoz/isoparser.git#egg=isoparser
